Uploading source code how HW09
I'm not sure of the implementation of the smooth shading. For scene 3 I believe I got the desirable results, however, I can't say the same thing for scene 4 and 5. I'm not exactly sure where the error might be. My last suspicion is the evaluation of the normal vector in the function where the if statatement smooth shading == 1 but still couldn't work out where exactly is and how to fix it. Something else that makes me wonder is why it would work properly for scene 3 and not for the rest